+++ Source/Core/GeometryPipeline.js
@@ -1187,20 +1187,13 @@ define([
         return geometry;
     };
 
-    /**
-     * DOC_TBA
-     */
-    GeometryPipeline.indexTriangles = function(geometry) {
+    function indexTriangles(geometry) {
         if (typeof geometry === 'undefined') {
             throw new DeveloperError('geometry is required.');
         }
 
         if (typeof geometry.indices !== 'undefined') {
             return geometry;
-        }
-
-        if (geometry.primitiveType !== PrimitiveType.TRIANGLES) {
-            throw new DeveloperError('geometry.primitiveType must be PrimitiveType.TRIANGLES.');
         }
 
         var numberOfVertices = Geometry.computeNumberOfVertices(geometry);
@@ -1219,18 +1212,11 @@ define([
 
         geometry.indices = indices;
         return geometry;
-    };
+    }
 
-    /**
-     * DOC_TBA
-     */
-    GeometryPipeline.indexTriangleFan = function(geometry) {
+    function indexTriangleFan(geometry) {
         if (typeof geometry === 'undefined') {
             throw new DeveloperError('geometry is required.');
-        }
-
-        if (geometry.primitiveType !== PrimitiveType.TRIANGLE_FAN) {
-            throw new DeveloperError('geometry.primitiveType must be PrimitiveType.TRIANGLE_FAN.');
         }
 
         var numberOfVertices = Geometry.computeNumberOfVertices(geometry);
@@ -1253,18 +1239,11 @@ define([
         geometry.indices = indices;
         geometry.primitiveType = PrimitiveType.TRIANGLES;
         return geometry;
-    };
+    }
 
-    /**
-     * DOC_TBA
-     */
-    GeometryPipeline.indexTriangleStrip = function(geometry) {
+    function indexTriangleStrip(geometry) {
         if (typeof geometry === 'undefined') {
             throw new DeveloperError('geometry is required.');
-        }
-
-        if (geometry.primitiveType !== PrimitiveType.TRIANGLE_STRIP) {
-            throw new DeveloperError('geometry.primitiveType must be PrimitiveType.TRIANGLE_STRIP.');
         }
 
         var numberOfVertices = Geometry.computeNumberOfVertices(geometry);
@@ -1299,22 +1278,15 @@ define([
         geometry.indices = indices;
         geometry.primitiveType = PrimitiveType.TRIANGLES;
         return geometry;
-    };
+    }
 
-    /**
-     * DOC_TBA
-     */
-    GeometryPipeline.indexLines = function(geometry) {
+    function indexLines(geometry) {
         if (typeof geometry === 'undefined') {
             throw new DeveloperError('undefined');
         }
 
         if (typeof geometry.indices !== 'undefined') {
             return geometry;
-        }
-
-        if (geometry.primitiveType !== PrimitiveType.LINES) {
-            throw new DeveloperError('geometry.primitiveType must be PrimitiveType.LINES.');
         }
 
         var numberOfVertices = Geometry.computeNumberOfVertices(geometry);
@@ -1333,18 +1305,11 @@ define([
 
         geometry.indices = indices;
         return geometry;
-    };
+    }
 
-    /**
-     * DOC_TBA
-     */
-    GeometryPipeline.indexLineStrip = function(geometry) {
+    function indexLineStrip(geometry) {
         if (typeof geometry === 'undefined') {
             throw new DeveloperError('geometry is required.');
-        }
-
-        if (geometry.primitiveType !== PrimitiveType.LINE_STRIP) {
-            throw new DeveloperError('geometry.primitiveType must be PrimitiveType.LINE_STRIP.');
         }
 
         var numberOfVertices = Geometry.computeNumberOfVertices(geometry);
@@ -1365,18 +1330,11 @@ define([
         geometry.indices = indices;
         geometry.primitiveType = PrimitiveType.LINES;
         return geometry;
-    };
+    }
 
-    /**
-     * DOC_TBA
-     */
-    GeometryPipeline.indexLineLoop = function(geometry) {
+    function indexLineLoop(geometry) {
         if (typeof geometry === 'undefined') {
             throw new DeveloperError('geometry is required.');
-        }
-
-        if (geometry.primitiveType !== PrimitiveType.LINE_LOOP) {
-            throw new DeveloperError('geometry.primitiveType must be PrimitiveType.LINE_LOOP.');
         }
 
         var numberOfVertices = Geometry.computeNumberOfVertices(geometry);
@@ -1401,7 +1359,26 @@ define([
         geometry.indices = indices;
         geometry.primitiveType = PrimitiveType.LINES;
         return geometry;
-    };
+    }
+
+    function indexPrimitive(geometry) {
+        switch (geometry.primitiveType) {
+        case PrimitiveType.TRIANGLE_FAN:
+            return indexTriangleFan(geometry);
+        case PrimitiveType.TRIANGLE_STRIP:
+            return indexTriangleStrip(geometry);
+        case PrimitiveType.TRIANGLES:
+            return indexTriangles(geometry);
+        case PrimitiveType.LINE_STRIP:
+            return indexLineStrip(geometry);
+        case PrimitiveType.LINE_LOOP:
+            return indexLineLoop(geometry);
+        case PrimitiveType.LINES:
+            return indexLines(geometry);
+        }
+
+        return geometry;
+    }
 
     function offsetPointFromXZPlane(p, isBehind) {
         if (Math.abs(p.y) < CesiumMath.EPSILON11){
@@ -1797,28 +1774,10 @@ define([
             }
         }
 
-        var primitiveType = geometry.primitiveType;
-        if (primitiveType === PrimitiveType.TRIANGLE_FAN) {
-            GeometryPipeline.indexTriangleFan(primitiveType);
+        indexPrimitive(geometry);
+        if (geometry.primitiveType === PrimitiveType.TRIANGLES) {
             wrapLongitudeTriangles(geometry);
-        } else if (primitiveType === PrimitiveType.TRIANGLE_STRIP) {
-            GeometryPipeline.indexTriangleStrip(geometry);
-            wrapLongitudeTriangles(geometry);
-        } else if (primitiveType === PrimitiveType.TRIANGLES) {
-            if (typeof geometry.indices === 'undefined') {
-                GeometryPipeline.indexTriangles(geometry);
-            }
-            wrapLongitudeTriangles(geometry);
-        } else if (primitiveType === PrimitiveType.LINE_STRIP) {
-            GeometryPipeline.indexLineStrip(geometry);
-            wrapLongitudeLines(geometry);
-        } else if (primitiveType === PrimitiveType.LINE_LOOP) {
-            GeometryPipeline.indexLineLoop(geometry);
-            wrapLongitudeLines(geometry);
-        } else if (primitiveType === PrimitiveType.LINES) {
-            if (typeof geometry.indices === 'undefined') {
-                GeometryPipeline.indexLines(geometry);
-            }
+        } else if (geometry.primitiveType === PrimitiveType.LINES) {
             wrapLongitudeLines(geometry);
         }